+++ src/commands/a350x/when.ts
@@ -1,9 +1,9 @@
 import { CommandDefinition } from '../index';
-import { CommandCategories } from '../../constants';
+import { Colors, CommandCategories } from '../../constants';
 import { createEmbed } from '../../lib/embed';
 
 export const EMBED = createEmbed({
-    color: 0x18b1ab,
+    color: Colors.DFD_CYAN,
     title: 'When is the A350X coming?',
     description:
         'Since this is a freeware project, we do not impose deadlines on our volunteer developers. ' +
