+++ src/lib/embed.ts
@@ -1,8 +1,8 @@
 import { EmbedBuilder, EmbedData } from 'discord.js';
-import { color } from '..';
+import { Colors } from '../constants';
 
 export const createEmbed = (options: EmbedData, timestamp?: boolean) => {
-    const embed = new EmbedBuilder({ color, ...options });
+    const embed = new EmbedBuilder({ color: Colors.DFD_CYAN, ...options });
     if (timestamp) embed.setTimestamp();
     return embed;
 };
