+++ src/logging/RoleLogs.ts
@@ -1,6 +1,7 @@
 import { createEmbed } from '../lib/embed';
-import { Colors, LogDefinition, getLogChannel, snakeToNorm } from '.';
+import { LogDefinition, getLogChannel, snakeToNorm } from '.';
 import { Role } from 'discord.js';
+import { Colors } from '../constants';
 
 export const roleCreate: LogDefinition = {
     event: 'roleCreate',
@@ -10,7 +11,7 @@ export const roleCreate: LogDefinition = {
 
         const embed = createEmbed(
             {
-                color: Colors.GREEN,
+                color: Colors.SUCCESS,
                 title: 'Role Created',
                 description: `**Role:** <@&${role.id}>\n**Name:** ${role.name}\n**Color:** ${role.hexColor.toUpperCase()}\n**Mentionable:** ${
                     role.mentionable
@@ -32,7 +33,7 @@ export const roleDelete: LogDefinition = {
 
         const embed = createEmbed(
             {
-                color: Colors.RED,
+                color: Colors.ERROR,
                 title: 'Role Deleted',
                 description: `**Name:** ${role.name}\n**Color:** ${role.hexColor.toUpperCase()}\n**Mentionable:** ${role.mentionable}`,
                 footer: { text: `Role ID: ${role.id}` },
@@ -53,7 +54,7 @@ export const roleUpdate: LogDefinition = {
         if (oldRole.color !== newRole.color) {
             const embed = createEmbed(
                 {
-                    color: Colors.ORANGE,
+                    color: Colors.WARNING,
                     title: 'Role Color Updated',
                     description: `**Role:** <@&${
                         oldRole.id
@@ -68,7 +69,7 @@ export const roleUpdate: LogDefinition = {
         if (oldRole.name !== newRole.name) {
             const embed = createEmbed(
                 {
-                    color: Colors.ORANGE,
+                    color: Colors.WARNING,
                     title: 'Role Name Updated',
                     description: `**Role:** <@&${oldRole.id}>\n**Before:** ${oldRole.name}\n**+After:** ${newRole.name}`,
                     footer: { text: `Role ID: ${oldRole.id}` },
@@ -81,7 +82,7 @@ export const roleUpdate: LogDefinition = {
         if (oldRole.mentionable !== newRole.mentionable) {
             const embed = createEmbed(
                 {
-                    color: Colors.ORANGE,
+                    color: Colors.WARNING,
                     title: 'Role Mentionable Flag Updated',
                     description: `**Role:** <@&${oldRole.id}>\n**Before:** ${oldRole.mentionable}\n**+After:** ${newRole.mentionable}`,
                     footer: { text: `Role ID: ${oldRole.id}` },
@@ -96,7 +97,7 @@ export const roleUpdate: LogDefinition = {
 
             const embed = createEmbed(
                 {
-                    color: Colors.ORANGE,
+                    color: Colors.WARNING,
                     title: 'Role Permission Updated',
                     description: `**Role:** <@&${oldRole.id}>\n**Removed:** ${snakeToNorm(perm)}`,
                     footer: { text: `Role ID: ${oldRole.id}` },
@@ -111,7 +112,7 @@ export const roleUpdate: LogDefinition = {
 
             const embed = createEmbed(
                 {
-                    color: Colors.ORANGE,
+                    color: Colors.WARNING,
                     title: 'Role Permission Updated',
                     description: `**Role:** <@&${oldRole.id}>\n**Added:** ${snakeToNorm(perm)}`,
                     footer: { text: `Role ID: ${oldRole.id}` },
